finish the basics of malloc

+++ include/types.h
@@ -3,17 +3,17 @@
 
 typedef signed char int8;
 typedef signed short int16;
-typedef signed int int32;
+typedef signed long int32;
 typedef signed long long int64;
 
 typedef unsigned char unsigned_int8;
 typedef unsigned short unsigned_int16;
-typedef unsigned int unsigned_int32;
+typedef unsigned long unsigned_int32;
 typedef unsigned long long unsigned_int64;
 
 typedef unsigned char uint8;
 typedef unsigned short uint16;
-typedef unsigned int uint32;
+typedef unsigned long uint32;
 typedef unsigned long long uint64;
 
 typedef _Bool bool;
